fix(audio): demo拆解与规范
